[od6.2] Few changes in uiCreatePicks
Summary: Removed the unit field. This leads to confusion. In time surveys users want to set a value in ms. In a feet survey, no one will want to set a depth value in meters.

Test Plan: Gui looks OK. Error message still works and picks or polygon are made with the correct z value.

Reviewers: prajjaval, raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12436