od: setting of VolProc::Step input distribution using IOPar
Summary: If a Step does not always have a fixed number of inputs, the creator of the step should have a way to set it dynamically using the Step::usePar/fillPar. Required for the fault thinning, which has 1, 2 or 3 datapacks input. For OD-1862.

Test Plan: used it for fault thinning. Tested Dip-Steering creation and Fault likelihood processing as well.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12429