od6.2: Restoring default write options and translator when a new dataset is selected
Summary: see workflow described in OD-1863. Purpose is to avoid inappropriate data storages (like 8-bit) when a new dataset is created the first time, even if the user has selected such dataset to set the output name as part of the new output name.

Test Plan: behaves as I want it. Switched several times the translator and/or the data storage parameters.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12465