od: Restored the user's LD_LIBRARY_PATH after OD initialization on Linux
Summary:
values given by the user to LD_LIBRARY_PATH were unset by init_dtect script initialization. As a result it was not possible to have OpendTect use the link to either Madagascar or GMT on Linux, as the required libraries may not be found if not hardcoded in the executables.
This had been done because of a specific issue with libGL, when the nvidia drivers are installed on Linux using the package manager.
Obsoletes D12470.

Test Plan: started OpendTect with various (and a missing) LD_LIBRARY_PATH on Farrukh's workstation which had such setting, and on other workstation where the driver was installed in the hard way.

Reviewers: bert

Reviewed By: bert

Subscribers: nageswara, raman, nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12472