[od6.2] Give axes fixed font
Summary:
Not sure if this is the right fix. But I'm also not sure whether we should let the user change the font of any axis. This at least fixes the problem of different font sizes in the well-tie module. The font of the log plots was already fixed, the graphicsscene axis now gets the same fontdata.
Alternative would be to create a new font type for axes to have it independent of any other text in graphics scenese.
JIRA: OD-1993

Test Plan: Changed 2D graphics font. Then opened well-tie module. Fonts are same now and everything is nicely aligned now.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12478