od6.2: protected the creation of the static TextTranslateMgr for threading issues
Summary:
if one creates a string like uiString str = tr("%1").arg( idx ) in the implementation of a paralleltask, and the manager was never accessed before, that is likely to trigger a crash because multiple threads will try to create the static manager. Thus added a necessary lock.
This will likely solved unexplained crashes at the start of the unconformity tracker execution when all seeds get inverted in separate threads.

Test Plan: no more crash.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12490