od6.2: added missing notification in uiPosSubSel
Summary: an external call to both uiPosSubSel::setInput functions did not trigger the required notification. As a result the change was not caught by the HorizonCube creator interface. Now the output HorizonCube sub-selection is really limited by the intersection of packages (horizons) ranges and the input SteeringCube range if the creator requires a SteeringCube input. Partly fixes OD-2089.

Test Plan: tested with HC processing where the horizon ranges are smaller than the input SteeringCube hsamp

Reviewers: bert

Reviewed By: bert

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12492