od6.2: Added a check for non-matching parentheses [OD-2127]
Summary:
Expressions with non-matching parentheses were parsed wrongly. Now I added a check on the whole expression before it goes for parsing.

Also, if the user entered an invalid expression, the whole text was wiped out and the user had to start all over again. This is also fixed now.

Test Plan: I get correct error messages for non-matching parentheses and the text remains there for me to edit. Also computed the attribute on an inline to confirm nothing is broken.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12544