[od6.2] Fix for picking faultsticks on random lines
Summary:
RandomLine ID was missing in datapack created for 2D viewer. After this fix, found a problem when switching between 2D and 3D viewer. Fixed that by making a similar condition as is done for picking in the 2D viewer.
Ported this code to 6.4 as well, but did not see a randomline id in the datapacks in the trunk. On purpose or missing?
JIRA: OD-2155

Test Plan: Add a random line. Display in 2D. Start picking fault sticks in 2D viewer. Sticks appear in both viewers. Picking can be done in both viewers.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12587