od6.2: Fix for 2D NN with trace step > 1
Summary:
For NN data extraction, there is no fixed GeomID. This was not handled well here. The reader reported a trace step of 1 whereas the line geometry and seismics had a step of 2. That led to a mess in the attribute engine.

This finally fixes Roar's problem.

Test Plan: I am able to train and apply the NN in https://jira.dgbes.com:8443/browse/OD-2142

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12598