od6.4: Added 2D Geometry import
Summary:
Support for import of 2D geometry. Need to do some more testing. Also working on importing SEG P1 format. This is not finished yet.
Please have a look at the current changes and see if the overall structure is OK, especially where the new geometries are added to the database.

Test Plan: Tested with TeapotDome which contains navigation files.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12618