od6.4: correct area measurement for 3D seismic datasets
Summary: the implementation was not based on the actual number of traces, i.e. the traces distribution was ignored. As a result the area reported in the seismics manager would always be far too large for irregular surveys.

Test Plan: tested in F3_Demo on a specific, very irregular dataset.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12622