od6.4: Fixed incorrect incidence angle from offset gathers not starting at zero
Summary: the computed angle gather was very wrong if the first offset trace of the gather was not at zero. To get it right tthe calculator needs to know if the Z range of the output gather is nmo corrected or not.

Test Plan: compared angle gather from synthetic gathers of offset range [0, 1000]m and [800, 1000]m, in a three layers case setup in SynthRock. The output value now match the values computed analytically for both interfaces.

Reviewers: nanne, raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12660