[od6.4] Fix for markers update
Summary:
When adding/removing markers to the 'Edit markers' dialog, the display is not updated. Fixed this by updating the list in the properties dlg and making sure all markers are always displayed when editing.
JIRA: OD-1267

Test Plan: Added and removed markers. Renamed and changed colors. When pressing 'Update' the markers in the 3D scene change. Also the list in the properties dialog updates.

Reviewers: raman, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12678