[od6.4] Fixed mousepos statusbar info
Summary:
When moving mouse over spectrum, wrong info was displayed in the statusbar. Also the readout stops when moving over the right-half of the graph.
Problem was specvals_ which was initialized with the wrong size.
Also renamed some variables in uiAmplSpectrum::valChgd function to make it more clear what's going on.

Test Plan: Values in statusbar look OK now.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12699