od: enable progress monitor for commands not launched in batch
Summary: was only enabled for CommandExecPars::RunInBG. Now it can also be used for CommandExecPars::Wait4Finish, providing that a log file name is provide

Test Plan: tested by applying a python script locally using the command launcher. For now I do not want it to be executed in batch.

Reviewers: bert

Reviewed By: bert

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12709