od: disabled open of uiStorableCollectionBuilder if nothing to open
Summary:
when the database does not contain any element of the required type, the open button should be disabled.
Button is restored when one item is saved successfully (could not imagin a way to do it fully internally, the user of the class needs to call a public function).

Test Plan: tested from machine learning GUI.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne, bert

Differential Revision: https://phabricator.dgbes.com:8443/D12719