[od6.4] Support server paths
Summary:
Windows server paths were not supported but requested by clients. It avoids mapping each and every network location to a drive. See description in the JIRA issue for more info.
JIRA: OD-2212

Test Plan: Tested setting the dataroot to a network location as well as importing a horizon ascii file. Both work, but there are probably other things to test as well.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12716