od6.4 : OD-2210, MMP handling Batch Host file on Windows
Summary: Earlier due to administator rights, host information file stored on netowrk could not be read. Now the od_batchhost.exe is launched on all platforms without any administrator rights. If file is not writable it does not allow you to add host too.

Test Plan: Tested. Above beaviour is seens. If file is in netowkr location it is read and saved properly

Reviewers: raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12729