[od6.4] Horizon-Fault intersection line does not always display
Summary:
Not sure what the real problem is. I found that the intersection is calculated correctly, but that some post-process fails because of the order of the points. As far as I understood the algorithm I concluded I had to change a false into a true to fix the problem.
JIRA: OD-2258

Test Plan: Problem is fixed in survey we received from client. Also tested in F3. There it always worked and it still works.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12755