od6.4: Replaced ParallelTask with Executor to fix randomness
Summary: While testing the fix for OD-2258 we found some random results which sometimes gave good intersection while some other times showed intersecting curves. This randomness was because of an unnecessary ParallelTask that was not contributing to the performance anyway. The Mutex made sure the whole operation was still single-threaded but with an unpredictable order. This eventually resulted in the intersection points being in a different order after each run.

Test Plan: Tested in OD-2258 test survey as well as F3_Demo. The results look good. Also printed the coordinates which are now always in the same order.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12763