od: solved few issues with OpendTect-HDF5 cmake setup
Summary: cmake does not set the same cmake variables before and after v3.10. Hence added an adapter. Also made sure that if a new HDF5_ROOT is provided, and the configuration (cache) has a different path, the new location is used to update the cache.

Test Plan: tested Linux, cmake 3.7.3 and 3.10.2, Windows: 3.7.3. Tested new configurations, and update of existing configurations.

Reviewers: raman

Reviewed By: raman

Subscribers: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D12772