od6.4: added possibility to use the virtual memory for mem allocation
Summary: providing that the swap memory is placed on an SSD, this fix allows using this available space for memory allocation. We cannot determine the composition of the swap memory, thus the feature would be enabled by an environment variable (and user setting?).

Test Plan: tested on Windows and Linux, the correct values are reported in the status bar, and memory allocation did not fail after exceeding the physical memory (but since my swap was on a HDD, the system got very very slow).

Reviewers: nanne

Reviewed By: nanne

Subscribers: bert, raman

Differential Revision: https://phabricator.dgbes.com:8443/D12787