od6.4: fix for saving of SynthRock model parameters
Summary: changes made to the model were not saved to disk unless the model was 'realized' by pressing the Go button. That originates from some model generators having an edited version, which was unknown of the save function.

Test Plan: tested Basic and Stochastic modules. Changes made to simulation units are now saved without having to press the Go button

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12788