[od6.4] Support re-start of volume builder process
Summary:
This was simply not implemented. Attribute processes always write individual inlines to disk, also when the number of inlines per job is >1. Added some code to make sure the detection of already processed lines works for the volume builder where larger chunks of data are written to disk.
Number of inlines per job is written in par file and can not be changed when restarting.
JIRA: OD-2285

Test Plan: Started a SteeringCube process, and let it fail for a few chunks. Did a restart to check if it correctly recognizes what to process.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12789