od6.4: fixed crash in prestack attribute
Summary: the class member PropCalc::Setup::useangle_ is wrongly used by the prestack attribute. As a result the attribute crashes when applying on an offset gather input, and the angle computation is switched off. The gather type itself can be angles. Thus added a conversion where the setup from the attribute is converted into a propcalc setup, and used the expected logic in PropCalc computer.

Test Plan:
added offset-based attributes to my existing attribute set for testing. Tested all previous attributes in addition to the new ones, and re-tested the SynthRock prestack attributes.
See OD-2194 for the testing data. Complements it by fixing OD-2286.

Reviewers: prajjaval

Reviewed By: prajjaval

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12793