od6.4: solved crash in rare cases
Summary: got a crash when building with O2, O3 optimization modes, with CXX_FLAG march=native. Solved it by moving the implementation of the destructor of FormulaSet to the cc file, and breaking the implementation in two statements.

Test Plan: no more crash when closing OD and switching to another survey.

Reviewers: nanne, bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12805