od: Made Isochron menu item dependent on Z axis type [OD-1005]
Summary: Changed the string "Isochron" to be dependent on either time or depth on Z-axis in a survey and change to Isochron for time and Isochore for depth.

Test Plan: Checked for a survey with z axis: time (F3_Demo_2015) and with survey with z axis: depth (Anadarko_Gunnison_Depth).

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12762