od: Fixed wrong redirection of od_cout in BatchPrograms
Summary:
od_cout() is implemented as od_ostream::logStream(). Batch programs do not have a log file, thus the implementation of OD::logMsgStrm() reaches the last if statement, where it was set to cerr instead of cout. As a result od_DBMan for instance did not send its response to stdout, and thus an empty string could not be parsed.
Fixes the python training error.

Test Plan: ran od_DBMan by redirecting both cout and cerr to two different files. Now the result is written to the correct stream (file)., and python dbman gets the correct answer.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12845