Use si's transform to calculate X/Y coordinate. OD-2321
Summary: To calculate the centerpos X/Y, the current SI() was used in stead of using the transform of si_.

Test Plan: In survey selection window, clicked on other survey with set CRS, open info window and go to the CRS settings and the coordinate transformation dialog. Default info looks OK now.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12876