OD: Make UI of Polygon properties consistent with UI of Polygon properties of the Basemap. Jira id:OD-2359
Summary:
OD: Make UI of Polygon properties consistent with UI of Polygon properties of the Basemap. Jira id:OD-2359
The Basemap and 3D scene had different properties for polygons. Those not in 3D scene have been added.

Test Plan: I have checked it on the Polygon Display Properties Window for proper display of UI as well as proper functioning of the properties.

Reviewers: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12929