od6.4: Display correct sampling rate in browse/edit option in seismic manager
Summary: When sampling rate is less than 1, the rows of the browse/edit option in the seismic manager appear as shown in fig OD2276_a.PNG This is because of the use of a macros that rounds up the input value.

Test Plan: Tested on a copied survey with sampling rate of 0.25. The result is shown in OD2276_a.PNG

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12940