od6.4: Display correct sampling rate in 2D lines in 2D seismic manager
Summary: When sampling rate is less than 1, z step in 2D line in the 2D seismic manager was being rounded off.

Test Plan:
Tested on 2D lines generated from the processing window in a volume with z step of 0.5 and 0.25.
The created lines now display the correct z step in the manager window (image attached)

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12949