od6.4: Incorrect default range in position dialog of a slice for z sampling step less than 1
Summary:
This is part of OD-2276
There were 2 issues:
[1] For z step < 1, the stop value of the z range in the position dialog was incorrect.
[2] The position dialog still showed z value range till 1 decimal even when z was an integer

Test Plan:
Tested on inl and crl slices with z step size of 4, 0.5, 0.25.
The correct z range corresponding to max range of survey is displayed in the position toolbar along with the same no. of decimals as in the z step.

Reviewers: nanne, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12948