od6.4: improving the user-interface to setup network communications
Summary: setting up a listening server to any host makes it by default listen with an IPv6 addresses. This is most often not desired, as we will prefer to listen by the name of the localhost (rather then "localhost"), conformable with IPv4 communication accross the LAN.

Test Plan: server is started according to the required settings. However the qtcpserver reports back the address "0.0.0.0" when being requested to listen from the localhost hostname. This is not yet compatible with the qtcpsocket, as this address "0.0.0.0" will not connect with the listening server. So we may want the server to keep the hostname with which it was started, to provide the correct authority code so the socket connect will work. Please advise.

Reviewers: nanne

Reviewed By: nanne

Subscribers: wayne, raman

Differential Revision: https://phabricator.dgbes.com:8443/D13038