Fix for uiTreeView::setEmpty()
Summary: uiTreeViewItem objects were not deleted in setEmpty(). Hence they were not removed from the static list odqtobjects_. This messed up the parent-child relation in the uiTreeView object when it was re-populated after a setEmpty() call.

Test Plan: Loading of new plugins in the plugin manager window does not mess up the selection anymore.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13129