OD-2363: When creating a survey from 2D data, it won't allow to proceed if CRS is selected
Summary:
[Make a new 2D survey by scanning a SEG-Y file.
Select any SEG-Y file, e.g. from Raw Data of F3.
In SEG-Y wizard pick CRS (ED50 / UTM zone 31N)
See that XY are set to 1e30.]
This issue has been solved. Basically, it happens only if the survey did not have a CRS selected yet and had a 'Unlocated X/Y' system as mentioned by Nanne in the comments on this Jira issue. It has been made sure that the option to select CRS on import is not available if that is the case.

Test Plan: test on all combination of 2D and 3D seismic data and inl/crl and X/Y

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13141