Title can be set as a custom title or automatically
Summary:
When saving the 2D viewer as an image, the window caption is not included. Added option to set title in viewer automatically, where title is as in caption, eg In-line 425.
JIRA: OD-2538

Test Plan: Opened 2D viewer for inl, crl, z-slice, 2D line and random line. Title looks good. And in case of inl, crl, z, the title updates after moving.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D13158