od: adapted uiRangePosProvGroup for multi-2D line case
Summary:
this class is the GUI for Seis::RangeSelData, which handles both 2D and 3D cases. Extended the GUI in such
a way that this uiGroup can also handle 2D as easily as 3D selections.
In most cases the number of lines to be used is already known and does not change, hence the line name selection
is a uiGenInput and not a uiListBox. That keeps the interface nice and compact.
To be used in machine learning GUI, but also interest for the crossplot selection.

Test Plan: tested from crossplot GUI.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13126