od: adapted error message for HighDPI screens
Summary:
The existing warning is not adapted to 4k screens setup with their windows
default scaling at 250%: the size returned by uiMain::getScreenSize is an effective
screen size, it is not the physical display size (hence it can go below the limit of 768 pixels.
The user can improve the apparence of OpendTect windows
by setting the windows display scaler to a lower value like 200%. I am not sure that we can
do much more.

I tried many possible Qt functions to get the pixel ratio, but most of them returned
a double value rounded to the nearest integer (2. for a scale set to 175%), instead of
an actual non-integer value.

Test Plan: have set to scale to 175% on a FullHD monitor to force getting the warning. Makes it relatively equivalent to 250% on a 4k monitor.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13263