Fixed forced correct detection of python environment
One can set OD_FORCE_PYTHON_ENV_OK to skip the tests run in
PythonAccess to ensure the environment is valid. Setting it to Yes
made it fail if the python source was not a system level, because
neither the activation nor the environment name were retrieved from
the settings and set in the appropriate class members.

Tested by launching the machine learning plugin with and without.
Also starts now with the environment variable set, providing that the
user settings are correct. Also tested that the content of the
*_requirement.txt file is ignored when this variable is set.