Fixed memory issues while filtering a SteeringCube
Summary: This is just a hack for beta release while I port the better implementation from trunk. Basically, the filter step was unaware of the nr components in the input SteeringCube. This resulted in system running out of memory when the filter step allocated memory for the output DataPack.

Test Plan:
Tested in debugger, memory estimate is correct now, when a Filter is applied
1. after creating a SteeringCube
2. While creating a SteeringCube.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13289