Fixed loading of new python toolbar [devel]
Toolbar was added twice when od_main was launched, and may crash
od_DeepLearning_CC
It should not be added in the constructor of uiSettingsMgr, because this
object will be created before the main window of the application is constructed.
Besides only uiODMain will need it, hence moved its initialization there.

No more pErrMsg when launching od_main. Toolbar appears at the start of the
application, and stays on survey change. No error when closing the application
either.