Further fixes in submission of commands to Qt [devel]
Splitting the command line string into program name and arguments.
Qt now fully takes care of adding the quotes to the command arguments, so
we no longer need to do it ourselves.
Allows advanced commands like in test_od_iostream to run successfully again,
even if the temporary file has a space in it.