Added variable OD_TMPDIR to set the path for temporary files
Allows the user to write all temporary files to another folder.
Can be used if the user cannot writable in the default temp folder
from the Operating System, or if using that path generates issues
(path is too long, contains spaces...)
Likewise, allows testing OpendTect to see if it can operate on a temporary
folder with a space in its name.