Fixed setting of PYTHONPATH environment variable [devel]
The PYTHONPATH is a merge of three sources:
1- User environment
2- OpendTect modules odpy, dgbpy
3- Custom PYTHONPATH saved in OpendTect settings

To allow changing the user setting (-3-), one needs to have
two BufferStrings saving the items 1, and 1,2. The second was missing,
thus added it in this commit.
Also PYTHONPATH should not change when the virtual environment changes,
only when the user settings change, from the Python settings dialog.