Fixed raising of foreign top level application windows
Added two notifiers windowShown, windowHidden
Moved two functions from uiDialog to uiMainWin
Fixed the setting of window flags such that it also works on finalised widgets.
This requires some wait time otherwise it does not work.

Renamed showOnTop to showAndActivate, since it is quicker to call
the activate function that to change twice the window flags.