Unified OD services under a same network server [devel]
One application uses a single network server for its communication,
regardless of whether it is od_main, od_DeepLearning_CC or any other

Class derived from either ODServiceBase of uiODService only have to
re-implement doParseAction and/or doParseRequest to handle their own
packets, or let the base class do it.

ODServiceBase is the base class for services which do not use UI components.
Example: batch processes (to be done) or the bokeh service manager

uiODService is the base class for services integrated to a uiMain application
Example: uiODServiceMgr for od_main, uiDeepLearningServiceMgr for od_DeepLearning_CC

uiODServiceMgr remains the only manager to register the external services.