Make detached subprocesses write stdout and stderr output to log file
Summary:
Would like stdout and stderr of mlapplyrun.py to be written to the
mother process log file. This seems to be the intent on the original code
but the method used doesn't work.
To Fix OD-2693 I want to have mlapplyrun announce the actual start of training
by writing to stdout and recognising that message in bokeh_machine_learning/main.py

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13354