Made odinst class work in development environments too [devel]
installerbasedir must be set to binary dir if both applications are in the same binary dir.
Made it work if OD/Installer applications are in built in Debug mode.
mRelRootDir should not be dependent on the application current path, I see no
reason not to set it to GetSoftwareDir for windows too.