Fixed crashes in well manager
Summary:
the function Well::Man::add in inherantly unsafe, as we do not
know what was loaded in the Well::Data object.
Luckily it is not used in our code, so marked it as deprecated, and tried
to make a good guess in case it is used by external plugins.
I also reviewed the ref/unRef calls, please mainly QC this part of the revision.

Test Plan:
started well data management table, loaded wells in the scene.
No crash during runtime or when closing od_main.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13398