Fixed blocked installer issue on Windows
File::isWritable is not sufficient on Windows to assess if a user member of the
admin group can write in a folder owned by the admin user (like in C:\Program Files).
Hence one must try to create a file inside the target folder to test its write status.

Re-cycled the canInstall function, which was not used anywhere (should be).

Fixes OD-2790