Fixed File::isWritable implementation for admin users on windows [devel]
If the folder is owned by admin, a UAC prompt is required to write in that folder,
even if the user is member of the admin group.
QFileInfo::isWritable may falsely return true, but creating a file stream and
checking if it is OK returns the correct answer.

One should see if it is possible to skip that test for a non-admin user,
if QFileInfo::isWritable() returns true.