System hangs when removing inline after scrolling from position window
Summary:
When removing an inline, it closes the position dlg which in turn closes
the scroll dlg. This results in a double lock.
unlock should be done before deleteAllChildren to fix the freeze.

Test Plan: Started automatic scrolling of an inline. Stopped this and removed the inline from the tree. No freeze.

Reviewers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D13435