Using Windows API for file writing permissions and AppLocker
replaces the creation of an output stream, which disturbs
the filemonitor

Improved AppLocker recognition
Some windows machines may still report that they are running the
Entreprise version of Windows, when they are not (because it joined
a standard network domain).
In that case AppLocker may work incorrectly if policy rules were previously
installed, are cached, and still partially used.

Fixes OD-2805 and OD-2806