Improve handling of depth domain data in frequency filter attribute
Summary:
Fixes OD-2574
Made attribute UI more responsive to z domain (depth vs time)
Also fixed spectrum displays issues. Previously filter was displayed in linear amplitude
and data spectrum in log amplitude. For consistency just using linear amplitude for both.
The slope parameter shown in the dialog is actually misleading. It is just the log ratio
of the start and end of the filter transition zone and doesn't include any information about
the filter amplitude response. Might be better to replace this with the % width of the taper.

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13462