Improved the test programs in Network [devel]
Test programs could fail when the port was not available, and used
a lot of old APIs. Updated the content by really using the
Authority object, and centralizing the command line parsing/MachineCommand
building inside that object.
The test programs should not also be capable of testing the local
server implementation with the flag --local