Add filtering by translator name or omf metadata keys to uiIOObjSelGrp and uiObjFileMan
Summary:
Prajjaval asked if I could port this feature from master to od6.6 Fixes OD-2825
I have also modified uiSeisFileMan as an example of its use for filtering by translator.
If you have a mixture of seismic data in CBVS, Blocks and SEGY in a survey this might
be helpful.

- Fix lint errors

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: prajjaval

Differential Revision: https://phabricator.dgbes.com:8443/D13471