Fixed bug in Math and Reference attrib. Showed wrong X/Y values.
Summary:
In both attributes it was calculating a SI().transform on the currentbid_.
For 2D this has a line/trace number and thus can't be transformed to a Coord
directly. Added new functions for this in Provider class.
Fixes OD-2832

Test Plan: Tested Math and Reference attrib in a 2D and 3D survey. Shows correct values.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D13480