Fixed communication between od_main and machine learning apps
Summary:
the wrong authorities are used to send requests, which thus never
reach the target application.
First I am porting some work done in the Batch module by Prajjaval, but doing
so I am cleaning up a lot the BatchProgram class. For now od_process_attrib
runs, but with the added even loop that we need for the network services.
Will be continued, as the connection between ODServiceBase and BatchProgram
is not yet implemented.

Test Plan: compiles so far.

Reviewers: nanne, arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D13488