now adding firewall rules
Summary:
Firewall rules were not even added when OD was launched with admin credentials, now forcing the command to be executed with admin rights should solve that issue.
Installer was also launching the window but rules were not added, this should fix it too.

Test Plan: Tested by launching OD from dev env and adding rules, but this is not a good test, since MSVC is running on admin credentials. Need to test it via installer.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D13498