Free memory on TypeSet erase
Summary:
Currently calling erase() on a TypeSet just calls std::vector.clear() which sets the size
to 0 but doesn't change the capacity (memory used) of the underlying container.
This change forces release of the memory by replacing the contents of the TypeSet with
an empty vector.

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13500