Added better tools to add contextual messages
Summary:
I want a tool that can easily be enabled with the environment
variable DTECT_DEBUG, for a certain level, and especially which does not pollute
the console output with lower level messages.
For instance in this example the developer of classes derived from
ServiceMgrBase should only get the corresponding messages at runtime,
but nothing from Basic, uiBase or others
About the "EXCLUSIVE" part of the mask, shall it be served with an extra
environment variable, or with an extra function argument?

Test Plan: only the message in ServiceMgrBase::useServer gets displayed

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D13503