Added required event loop to test_oscommand program [devel]
The test program was issuing an PE message that indicated that a timer
was used without event loop. This rework provides an event loop to the test program

Also this test program will always fail is DTECT_DEBUG is set, thus locally unsetting this
environment variable.

Finally this test program has the actual main implementation in a function testMain(int,char**),
which ensures that all local variables are deleted before the call to ExitProgram. This avoids
the need to declare the testing class using a PtrMan, and have to delete it after the event
loop execution is finished.