Fixed issues with services with survey change and application closing
Too many close signals were triggered, and/or applied in the wrong order.
Now the external services behave well upon closing the client application
or switching the survey from the client application.

Because of ABI constraints, some of the implementations are duplicated in the classes
derived from ServiceServerMgr. They should belong to this base class.